Разделение js кода на модули, импорт в HTML
